pyproject.toml
[tool.poetry]
name = "SQLAlchemy-MissingGreenletError"
version = "0.1.0"
description = "MissingGreenletError"
authors = ["Palaniyappan Ramanathan <vickypalaniappan12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.30"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
